Fix use of getWeights in setting weights for next iteration.
getWeights() is used in two contexts: first, conversion of moments
to weights; and second, conversion of differences in weights to weights
for the next iteration (under the "product of gaussians" assumption).
I'm finding cases of two blended objects (so a deblender would certainly
help, but I think we shouldn't need one for this code to be robust)
where the determinant is negative in the second context.  I don't
believe that the "careful" handling of negative determinants is valid
in that case (the iteration has failed because the assumption has been
violated; the code handles this by going on to calculate unweighted
moments), and so I've turned it off for that context only.