After detecting sources, optionally re-estimate the background.
When we're re-estimating background we need to run a negative detection
step so as not to bias the sky level; this mask plane is cleared after use.

N.b. The footprint grow is NOT applied to the objects being measured.
This is probably a bug due to a previous round of refactoring, but I'm not
fixing it now as it'll cause more blending and we're not (yet) running a
deblender

Also pass in a Log object, and configure if the Footprint grow is isotropic