Test the objectSizeStarSelector
N.b. I had to fiddle with the image generation to make the tests pass,
and with the secondMomentStarSelector configs to get it to work --- it was
a miracle that it ever did for these sparse images with Ixx \propto Iyy

N.b. relaxed the tolerance for extrapolating the PSF in the sub-image test,
but also added a more stringent smaller-extrapolation case