No need for this swiggery
Jim says, "It was one of many things I was trying at some point to convince swig that it was not an abstract class, because it used to be confused about that.  Either I found some other way to fix it and didn't remove this, or swig just got smarter at some point.  I don't recall which."