Rework silly centroid test with control parameter.
The silly centroid code serves as an example for new algorithms, so it's
useful to see how it looks when there's a control parameter.  Also expanded
the test to show the use of the silly centroid within the measurement Task.

Changes are purely to the tests, not to anything functional.