Make convergence criteria for adaptive moments configurable (#2638).
In calibrating an image, a first measurement pass is made before cosmic
rays are excised (because we don't know the true PSF yet, and attempts to
flag CRs before we know the PSF will cause us to flag stars which will
adversely impact our ability to determine the PSF). Measurement algorithms
therefore need to be able to cope with CRs, and in the CFHT-Megacam u-band
we get LOTS. This has revealed that the SDSS adaptive moments code is
rather slow on CRs.

The code contained three #define-ed constants: MAXIT, TOL1, TOL2. Backing
off on these allows the SDSS shape measurement to run faster on Megacam
u-band data.  Make these configurable through the SdssShapeControl.

There is a minor API change to the getAdaptiveMoments function, but
because of the use of default arguments, this should be
backwards-compatible.