Remove ShapeletPsf; it's too bitrotted to bother updating.
I also know Mike Jarvis has a better version elsewhere that he has
been keeping up to date, and if we ever want ShapeletPsf back, we
should re-port that.

I have not removed everything related to the ShapeletPsf code, because
some of it may be of more general use
(e.g. sizeMagnitudeStarSelector).