CoaddPsf, WarpedPsf: optimise warping execution time by 15x
Picking some low-hanging fruit to speed up use of CoaddPsf when there
are lots of inputs.

97% of 1818 sec was being spent in warping.  Two ways I found to speed
this up by an order of magnitude:
 * Use a less expensive warping kernel than lanczos5 (lanczos3 by default)
 * Cache the warping kernels in the CoaddPsf (10000 samples by default)
I also tried converting the images being warped to float and warping
those, but that resulted in a net increase in execution time.

The default warping kernel and cache sizes are configurable in the Ctor.

After the optimisation, the execution time is 116 sec.