PsfCandidate: improve masking of unwanted pixels (#2933)
Sources neighbouring PsfCandidates are one of the principal annoyances in
PSF determination, as they can use principal components that might
otherwise be used to model real PSF variation.  We already mask out
detections within a PsfCandidate that are not connected to the candidate,
but neighbours can sometimes be close enough for the footprints to overlap.
Also, there may be sources fainter than the (high) threshold used in the
first-pass detection lurking within the PsfCandidate, and the current code
does not deal with them.

Here are two features which help improve the masking:
1. Pixel masking for blended sources (with more than one peak): mask pixels
in the footprint that are not closest to the central peak.  This new
feature is always active, as blends are a clear contaminant.
2. Pixel masking for external sources: mask pixels above some threshold
except for pixels in the footprint.  This catches faint sources surrounding
the candidate.  This feature is not activated by default because it
requires an accurate variance plane.

With both of these features, I get much better PSF determination on dirty
HST images (lots of leftover CRs).