GaussianFlux: fail when forced and shape failed
If we're running forced and the shape measurement failed, then
the shapes could be NAN, and we might hit asserts in SdssShape.cc's
getAdaptiveMoments().  Better to fail gracefully.  The particular
flag to check is configurable (well, its suffix is, to prevent
mismatch between the sources of the shapes and the flag).