Reset background defaults back to the original ones
The intention is to eventually set useApprox=True (i.e. Chebychev
Approximation) for background estimation.  However, in looking into
the relevant code in afw/math, there is some clean-up and restructuring
that needs to be done before resetting the defaults (which may also
require adjusting some defaults in the calibrate stage to be more
appropriate for the approximation, as opposed to inperpolation,
scheme).  This is the subject of a new ticket DM-2920.  Ticket
DM-2921 is to set the defaults to useApprox=True.