CoaddPsf constructor: use "goodpix" field, if available
Modify the CoaddPsf constructor to use the "goodpix" field
of "catalog", if present, when computing _averagePosition.
The only required change was to copy the "goodpix" field,
if available, when copying the catalog. The existing
computeAveragePosition method already used the information
if it could find it, and silently ignored it otherwise.