Interp: Remove unused Psf argument from interpolateOverDefects
While this may have been added with the intention of making use of a
Psf in the defects interpolation in the future, having it there and
as a required argument to interpolateOverDefects but unused in the
function is misleading.  It also leads to many calls to this function
having to set a guessed/approximate psf for images that do not have
one associated with them.