Only take median position when points exist.
Previously, when we attempted to take the median position of a list of points we
relied on the (old, pre-1.10) NumPy behaviour of providing NaN as the median
of an empty array. Newer NumPy will raise an IndexError instead. Here, we
explicitly check that we have at least one point before taking the median.