Add joinMatchListWithCatalog to LoadReferenceObjectsTask
Moving joinMatchListWithCatalog funcion from meas_astrom's
AnetBasicAstrometryTask to be a method on meas_algorithms's
LoadReferenceObjectsTask.  This removes the requirement to go
through an instance of AnetBasicAstrometryTask to get to the
function.  joinMatchListWithCatalog only requires a reference
loader, so is a natural fit for LoadReferenceObjectsTask.  This
will also allow it to work with any kind of reference catalog
(i.e. other than a_net).