Update bit-rotted test
This test was disabled because it was not working. This caused some
of the interface to bit-rot. Updated to using functor keys for things
like position. Additionally fixed the test, supplying all information
necessary to do basic processing of the fake source. The test was
previously failing because it attempted to read properties from the
source object that were NaN, causing error handling to be prematurely
executed.