Make failure to measure aperture correction an exception.
The previous behavior - just warning - was clearly not sufficient to
catch a long trail of bugs and sorrow.  And this should be no more
fatal to well-configured processing on low-quality data than the
existing requirement that we have enough stars to model the PSF.

This required rewriting the aperture correction unit tests, as these
previously caused these silent failures while testing other aspects
of the system.