Re-direct all warnings to debug log
As part of the logic of the k-means algorithm, comparisons are made
between numbers and nans which causes numpy to issue warnings. This
commit captures the warnings using the warnings.catch_warnings context
manager and redirects them to a debug log at debug level 2 (rather than
say 1) because the warnings are relatively benign.