Added initial astrometry source selector, base class and tests.
This is the first non-PSF source selector, with a new base class for other
generic catalog-based source selectors. It was lifted from measOptimisticB for
use by jointcal.

Although it passes all tests and should work as a replacement for SourceInfo
from matchOptimisticB, it appears to be not quite strict enough for jointcal
yet, so may require some more work for that.