Fixed HtmIndexerTask's access of ctrCoord
HtmIndexerTask used ctrCoord.getRa and getDec, which are not
reliably present, depending on whether the object is passed
in as a Coord or a subclass (even if ctrCoord is actually a pointer
to an Fk5Coord or IcrsCoord, it may show up as a pointer to Coord).
So use getLongitude and getLatitude, instead.