starSelector: remove conditional setting of default width
The width was only being set if the width was currently zero.
This produced the following pattern:
* Processing image 1:
  * star selector: width is zero --> set to 21
  * psf determiner: set width to 41
* Processing image 2:
  * star selector: width is 41 --> leave
...

The lack of consistency between the width used in the first star
selector and the width used in subsequent star selector calls can
affect the choice of PSF candidates, and makes the pipeline results
subject to when it's run and in what manner.

To fix this, we always set the default width in the star selector.