Avoid astropy.Table to limit import clutter.
Importing astropy.Table brings in lots of dependencies like IPython, which get
saved to our config files. This is particularly a problem because the config
file for IngestIndexedReferenceCatalog is loaded each time the catalog is used,
and it tries to import astropy.io.hdf5 which crashes if it is not imported in
the correct order (See Astropy #4750, #1402). It is easier to avoid
astropy.Table entirely, and the column rename can be easily accomplished without
it.