Include catalog overplotting in 'if display' guard.
This moves the call to psfHist.insert() into a separate loop in
which we don't do any plotting (though we do create an array of
colors), since we want to run that regardless.

Before DM-7848, calling ds9.Buffering() with no display setup
was apparently a harmless no-op.  Even if the intent is that
it should be, and it is made a no-op again, I still think this
fix should stand; it's much cleaner to separate the algorithm
logic from the display logic.