Add "resized" method to Gaussian Psfs.
PSF-matching a WarpedPsf to a model PSF requires the ability to change
the kernel dimensions of model PSFs at runtime depending on the maximum
(or minimum) dimensions of a WarpedPsf across an image.  Because PSFs
are immutable, resized() produces a copy of itself with same model
parameters and specified new dimensions.