Add unittest for base_SdssShape_psf measurement
This test is being added here in meas_algorithms rather than meas_base
(where SdssShape lives) due to the need to apply a spatially varying
PSF model such that the PSF is different at each position in the image.
This varying psf model is built with meas_algorithms' PcaPsf (which is
not accessible from meas_base).