CoaddPsf: add more exception info
A common problem when calling CoaddPsf.computePsfImage() is that
WarpedPsf.computeBBoxFromTransform throws a RangeError stating
'Unexpectedly large transform passed to WarpedPsf'. This is
usually because the astrometry for an input is poor, but it would
be nice to identify *which* input: so add the exposure ID to the
exception message.