add ScienceSourceSelectorTask,ReferenceSourceSelectorTask
These will select sources for photoCal. That job is currently done by some
long function in lsst.pipe.tasks.PhotoCalTask. Moving it into source
selectors allows for code re-use.

The selections done here are very modular, using Config to configure the
limits and then giving them an 'apply' method to do the selection. This
makes it very easy to expand the selections or to write new combinations
of selections.