Update a unit test note and shrink an atol value
testPersistence in test_coaddBoundedField.py wants to check that
the unpersisted bounded field is identical to the original.
Unfortunately that is not the case, even after fixing DM-12270,
so I filed a new ticket and updated the comment accordingly.
I also shrunk the atol a bit (leaving more than 10x headroom
as margin for other compilers).