Replace appended result catalog with selected catalog (much faster).
The original version created a new catalog and appended one row at a time.
This was unnecessary as the catalog[selected] call already makes a suitable
catalog for return.  There is already no guarantee that the returned catalog be
contiguous in memory (and the old version was also not contiguous).