Update refcat loaders to use nJy fluxes
Update units in docstring and makeMinimalSchema flux units.

Add format_version to Indexed DatasetConfig, and use setDefaults to override
it when Ingesting new Indexed refcats.

Add functions to check for and convert old refcat fluxes.
Convert old refcats when they are read in, and issue a warning about them.

Convert afw flux->mag code to astropy where possible, and multiply by 1e-9
for the fluxErr/magErr code that will be removed in DM-16903.