Make refcat ingestion use multiprocessing and improve tests
Split the parallelized work into a separate module, so that it doesn't deal
with pickling a Butler, or DataRefs.

Add progress indicator to ingest manager.

The previous test code did not check that the ingested catalog actually
contained the same things as the input catalog. I've improved the tests
to do that, and also pulled out some shared code so that I could write
unit tests of parts of IngestIndexedReferenceTask and a separate test
of the multiprocessing portion of the code. That new test is run outside of
pytest, so that it doesn't interact with the pytest multiprocessing.