Special case classes for Gaia DR2 ingestion
Gaia's fluxes and flux errors cannot be trivially converted to nJy, so
I've subclassed the manager and ingester to handle that. It somehow
feels a bit hacky, but was not as hacky as I'd feared.

Update the "how to" document to note this change and use the exact
Config and script I used for the final ingestion.