Add BrightStarStamps class.
This new dataset class will allow for small image cutouts ("stamps"),
each centered around an individual bright star, to be saved to disk.
These can then be used to stack a large number of such stars to build
an extended PSF model. Each stamp image is normalized using the flux
of the stars in annulus around their center (since the central area
are saturated and/or contain numerous ghosts). The radii used to define
the annulus are stored in the metadata, along with each individual
annularFlux.