Fix SourceDetectionTask change of input exposure.
In the doSmooth=False case, the new exposure within
SourceDetectionTask shared the same pixel memory as the input
exposure. These pixels are modified in the task which appeared in the
input exposure as well; this was unwanted behavior.

Note that the more frequently used doSmooth=True did not have this
bug. It allocates new space for the smoothing convolution operation.