Fix up passing of epoch parameter among functions
The epoch parameter was not being passed consistently among the different
functions that take it as an input parameter.  The epoch is used to correct
the reference catalogs for proper motion and parallax (when present).  This
correction was thus not taking effect when the passing down of the parameter
was incomplete.  This situation is remedied here.

The epoch object being passed around is consistently of type astropy.time.Time
or None.  The value desired for the metadata "EPOCH" header set in the
getMetadataCircle() and getMetadataBox() functions is the float representation
of MJD.  This is obtained via the .mdj attribute of the astropy.time.Time
object.

This also adds a unittest to check that the proper motion correction is
occurring when loading a reference catalog via the loadPixelBox() funtion
The check here uses an extreme value for the epoch such that the differences
will be significant and reasnoably tested to have changed via
assertFloatsNotEqual.  Note that this simply checks that the coordinates do
indeed change when the epoch is passed to loadPixelBox().  It makes no attempt
at assessing the correctness of the change.  This is left to the explicit
testProperMotion() test below.