Ensure ApCorr fields are added deterministically
Sort the set returned by getApCorrNameSet() so that the fields named there
are added in a deterministic order.

This also adds an "attempt" to test this, which is difficult since the order
could end up effectively sorted accidentally, but we do our best here by
forcing the two test names to be originally added in an anti-sorted order.