Pre-filter loaded reference catalogs
The raw WCS available here for sky <--> pixel conversions is only reliable
for an area around the origin that is quite local.  On the other hand, a
given reference catalog shard that overlaps the (padded) CCD region may
extend well beyond the CCD region.  As such, attempting to convert from
sky -> pixel coordinates for the reference catalog (in order to assess
whether a given object lies within our padded bounding box of interest) can
lead to extreme extrapolation errors, occasionally placing objects inside
this bbox that, in truth, lie nowhere near the region of the CCD of interest.
This pre-trimming is based on the reference catalog coords and excludes
those lying outside a sky region defined by adding an extra padding to our
padded-bbox.  This should ensure a large enough projected area on the sky
that accounts for any projection/distortion issues, but small enough to
filter out loaded reference objects that lie well beyond the projected
detector of interest (and hence would have been vulnerable to WCS
extrapolation errors).