Refactor refcat ingest to base class for future gen3 work
Factor out butler.get/butler.put calls, so that we can use the rest of
the refcat ingest code for both the gen2/gen3 versions.

Replace `createIndexedCatalog` with `run`.

Name change ingest->convert, as the gen3 code only does conversion, not
ingestion.

Make refcat convert manager a config option:
* This eliminates the need to have a separate IngestGaia Task and lets us
write configs for gen3.
* Rename ingestIndexManager->convertRefcatManager to better reflect what
it is for (it is a purely internal class).

Cleanup/refactor convert manager to output better docs.