Add RequireSkyPosition class to science selector
We are now allowing source catalogs to be persisted even in light of
an astrometric fail in single frame processing.  Such catalogs will
have their sky coordinate (typically columns: "coord_ra" & "coord_dec")
values set to NaN. It is thus useful to add a clause that will select
against sources for which their sky coordinate entries are non-finite.
This clause is added explicitly to the ScienceSourceSelectorTask (but
defaults to False so will not change current behavior).

A unittest is also added to test_sourceSelector.py here.