Add WCS is None checks on dynamicDetection
The dynamic detection task makes use of forced photometry for measuring
"sky" sources, and this requires a WCS.  However, in principle, the
dynamic detection is not reliant on a true pixel-to-sky WCS so, if the
exposure provided has its WCS set to None, add in a dummy WCS for the
purpose of this task.

Note that we currently only make use of DynamicDetectionTask in the
context of multiband measurements on coadds, for which the exposures
provided will always have a valid WCS attached.  However, in principle,
this taks could be retargted in other contexts, so this check is
provided for that possibility.