Remove unhelpful script
The check script only checked that sources existed, it didn't test any of the
relevant conversion values (which is much harder), and it would be difficult
to update it for gen3. I'm removing it because it would probably just confuse
people and provide a false sense of security were it updated to gen3.