Update min S/N default on objectSizeStarSelector
The threshold set here should take into consideration the detection
thresholds in play for the catalog most commonly of interest here, namely
the one measured in the CharaterizeImageTask, which has:

detection.thresholdValue = 5.0
detection.includeThresholdMultiplier = 10.0

The value of 20 is too low for such a high detection threshold.  This low
value is not only misleading as to the range actually accessible, but also
allows for the occasional lower-than-expected S/N sources to pass the
thresholding tests for consideration in the PSF modeling (noting that this
effect will generally be insignificant, according to a detailed look at a
number of cases, but is still worth guarding against).