Remove defaultFilter and make filterName required arg
Removing defaultFilter means that we always need to specify a filterName when
loading a refcat, building metadata, or manipulating flux field schemas.
As far as I could tell, the tests here were the only place where filterName
was sometimes taken to be None, so this change from a kwarg to an arg should
be completely safe (and also the defaultFilter code itself didn't work the way
most people generally assumed).