Compute peak significance values in detection
Cleanup test_detection and test_negative so that I can test the new
significance calculations there.

Ensure that the threshold values are passed back up after local background
correction, so that we can calculate the significance at the end.

Use schemaMapper for significance: This means we don't have to modify the afw
footprint/peakRecord persistence code or deal with file compatibility.