Make WarpedPsf delegate to source PSF's computeImage.
Instead of warping the result of a computeKernelImage call with a
transform that does not shift at all, we now warp the result of a call
to computeImage with a transform that includes a shift to (0, 0).
When the source PSF has a custom computeImge implementation that can
do better than just Lanczos-shift its own computeKernelImage result,
this make the warping operation more like what stars on the
corresponding images experience.