Force (0, 0) for center of warped PSF images.
Previous logic forced the image size to be odd, but picked the center
of the kernel image bbox based on center of the transformed kernel
image bbox.  But the output kernel image center is an input: (0, 0).

This manifested as even-sizes images later in CoaddPsf; the union of
two odd-sized boxes with the same dimensions but offset center is
an even-sized box.