Final ingest->convert gen2->gen3 conversion/renaming
The useful methods in "ingestIndex" now live in
convertReferenceCatalog.py, which is where they are used. "ingest" in
gen3 is an entirely separate process.

As part of this, I removed the convert base class, because it was only
there for gen2/gen3 compatibility. I refactored the interface as part
of that, to remove methods that only were split to modularize the
gen2/gen3 differences.