Remove default ref_dataset_name
This was ticketed for DM-31817, but it makes sense to do it here.
Have to add that field back to the `bad_config.py` test, as it is still
a necessary field (so that the name of the refcat is tracked in
something other than the path it is written to).