Add position covariance to refcat schema
This removes the function `ReferenceObjectLoader.makeMinimalSchema` and
adds the free function `_makeSchema` to `convertReferenceCatalog.py`.
The new function adds the possibility to include 5-d covariance in the
reference catalog. This option is turned on using the
`ConvertRefererenceCatalogConfig` option `full_position_information`.

Update refcat schema docs to reflect the correct covariance matrix names.