Improve mask setting for dynamic detection
In particular, add an initial bright object detection pass with significant
growing of footprints to account for wings.  The negative polarity detections
in this pass help in masking highly over-subtracted regions.  In the next
"preliminary" detection pass, the negative polarity threshold is increased
(relative to the positive threshold) in order to avoid aggressive negative
detections that leads to severe overmasking of the image (leaving no room
for sky object placement).

This also adds the INTRP and SAT mask planes to the list to avoid in sky
object placement.

A few config options are added related to the above and the minNumSources
config is changed to be a minimum fraction of the total sky sources
requested.  The test is updated to reflect this rebranded config.