Improve Sky Object Placement
This commit switches sky object placement from pseudo-random to
quasi-random positioning using the Halton Sequence. During testing, the
dynamic detection task unit test returned failures for all tests which
employ sky objects. This is believed to be an inherent weakness within
the dynamic detection task highlighted by the changes made on this
ticket, and not related to any of the changes made here. As a
consequence, the relative tolerance for acceptance of those failing unit
tests was increased from 0.1 to 0.15 in order to allow this ticket to
pass at this time.