Force stamps to be ImageF
This allows guider files of type ImageI to be read in.
This does not solve the problem of ImageD stamps though.

The Stamps abstract class doesn't force MaskedImageF but
the main Stamps definition does.

If Guider files declared a different type of Stamps
class rather than the default then we could relax
this constraint on the reader.