Remove INTRP from avoid list in dynamic detection
As the number of input visits increases, the number of pixels with the
INTRP mask bit set increases, which eventually leads to a very large
fraction pixels with this mask bit set.  Thus, if INTRP is included in
the list of masks to avoid when laying down sky objects, very deep
coadds will end up with little-to-no available locations for sky
objects, leading to a failure in the dynamic detection.  Thus, the
safest thing to do is leave it out of the masks to avoid list.