Add and use information about fraction of good pix
In the dynamic detection task, compute and log the number of "good"
pixels in the exposure, where "good" is defined simply as not having
either the NO_DATA or BAD mask bit set.  If this fraction is greater
than 20% of the total number of pixels, yet we still fail to locate
enough spots to lay down sky objects, raise an RuntimeError exception.
The message includes the fraction of the "good" pixels marked either
DETECTED or DETECTED_NEGATIVE as this is a likely reason for this
situation.