Add method to get colorterm from ReferenceObjectLoader.
This isn't yet in an ideal state, because due to package-dependency
constraints it can only return the simple pydantic model in obs_base,
not the featureful thing in pipe_tasks, but those have the same fields
and can be converted.  Eventually I envision us moving more
functionality down to obs_base, and probably deprecating the pipe_tasks
version.