Add condition on number of peaks per footprint
If the detection mask ends up as a single, or few many-peaked footprints,
then all subsequent processing will fall over.  Impose a limit to the
maximum number of peaks in a given footprint for the detection mask.  If
this condition is not met, iteravely increase the detection threshold until
it does.  This is meant to be a no-op for most "typical" scenes and observing
conditions, but can help avoid a total downstream meltdown in certain cases
(e.g. very crowded regions).